feat(listbox): convert to ts
