+++ types/global.d.ts
@@ -1,0 +1,9 @@
+declare module '@neovici/cosmoz-dropdown/src/use-position.js' {
+	export function usePosition(opts: {
+		anchor?: HTMLElement;
+		host?: HTMLElement;
+		confinement?: unknown;
+		placement?: unknown;
+		limit: boolean;
+	}): void;
+}
